+++ FixedspawnDissonance/code/Artifacts/KinBossDropsForEnemies.cs
@@ -10,121 +10,83 @@ namespace VanillaArtifactsPlus
  
         public static void Start()
         {
-            SerializablePickupIndex BossItemWisp = new SerializablePickupIndex() { pickupName = "ItemIndex.SprintWisp" };
-            //SerializablePickupIndex BossItemVagrant = new SerializablePickupIndex() { pickupName = "ItemIndex.NovaOnLowHealth" };
-            //SerializablePickupIndex BossItemTitan = new SerializablePickupIndex() { pickupName = "ItemIndex.Knurl" };
-            SerializablePickupIndex BossItemBeetle = new SerializablePickupIndex() { pickupName = "ItemIndex.BeetleGland" };
-            //SerializablePickupIndex BossItemImp = new SerializablePickupIndex() { pickupName = "ItemIndex.BleedOnHitAndExplode" };
-            SerializablePickupIndex BossItemRoboBall = new SerializablePickupIndex() { pickupName = "ItemIndex.RoboBallBuddy" };
-            //SerializablePickupIndex BossItemParent = new SerializablePickupIndex() { pickupName = "ItemIndex.ParentEgg" };
-            //SerializablePickupIndex BossItemClay = new SerializablePickupIndex() { pickupName = "ItemIndex.SiphonOnLowHealth" };
-            SerializablePickupIndex BossItemMagma = new SerializablePickupIndex() { pickupName = "ItemIndex.FireballsOnHit" };
-            //SerializablePickupIndex BossItemElectric = new SerializablePickupIndex() { pickupName = "ItemIndex.LightningStrikeOnHit" };
-            SerializablePickupIndex BossItemShinyPearl = new SerializablePickupIndex() { pickupName = "ItemIndex.ShinyPearl" };
-            //SerializablePickupIndex BossItemPearl = new SerializablePickupIndex() { pickupName = "ItemIndex.Pearl" };
-
-            //SerializablePickupIndex BossItemLamp = new SerializablePickupIndex() { pickupName = "ItemIndex.MinorConstructOnKill" };
-            SerializablePickupIndex BossItemVoid = new SerializablePickupIndex() { pickupName = "ItemIndex.VoidMegaCrabItem" };
-
-
-
-            DeathRewards deathRewards = LegacyResourcesAPI.Load<GameObject>("prefabs/characterbodies/GolemBody").GetComponent<DeathRewards>();
-            deathRewards.bossPickup = new SerializablePickupIndex() { pickupName = "ItemIndex.Knurl" };
-            deathRewards = Addressables.LoadAssetAsync<GameObject>(key: "RoR2/DLC2/Halcyonite/HalcyoniteBody.prefab").WaitForCompletion().GetComponent<DeathRewards>();
-            deathRewards.bossPickup = new SerializablePickupIndex() { pickupName = "ItemIndex.Knurl" };
-            #region Beetle
-            deathRewards = LegacyResourcesAPI.Load<GameObject>("prefabs/characterbodies/BeetleBody").GetComponent<DeathRewards>();
-            deathRewards.bossPickup = BossItemBeetle;
-            deathRewards = LegacyResourcesAPI.Load<GameObject>("prefabs/characterbodies/BeetleGuardBody").GetComponent<DeathRewards>();
-            deathRewards.bossPickup = BossItemBeetle;
-            #endregion
-            #region Jellyfish
-            deathRewards = LegacyResourcesAPI.Load<GameObject>("prefabs/characterbodies/JellyfishBody").GetComponent<DeathRewards>();
-            deathRewards.bossPickup = new SerializablePickupIndex() { pickupName = "ItemIndex.NovaOnLowHealth" };
-            #endregion
-            deathRewards = LegacyResourcesAPI.Load<GameObject>("prefabs/characterbodies/WispBody").GetComponent<DeathRewards>();
-            deathRewards.bossPickup = BossItemWisp;
-            deathRewards = LegacyResourcesAPI.Load<GameObject>("prefabs/characterbodies/GreaterWispBody").GetComponent<DeathRewards>();
-            deathRewards.bossPickup = BossItemWisp;
-            deathRewards = LegacyResourcesAPI.Load<GameObject>("Prefabs/CharacterBodies/ArchWispBody").GetComponent<DeathRewards>();
-            if (deathRewards)
-            {
-                deathRewards.bossPickup = BossItemWisp;
-            }       
-
-            deathRewards = LegacyResourcesAPI.Load<GameObject>("prefabs/characterbodies/LemurianBody").GetComponent<DeathRewards>();
-            deathRewards.bossPickup = BossItemMagma;
-            deathRewards = LegacyResourcesAPI.Load<GameObject>("prefabs/characterbodies/LemurianBruiserBody").GetComponent<DeathRewards>();
-            deathRewards.bossPickup = BossItemMagma;
-            deathRewards = Addressables.LoadAssetAsync<GameObject>(key: "RoR2/DLC2/Scorchling/ScorchlingBody.prefab").WaitForCompletion().GetComponent<DeathRewards>();
-            deathRewards.bossPickup = BossItemMagma;
-
-
-            deathRewards = LegacyResourcesAPI.Load<GameObject>("prefabs/characterbodies/ClayBruiserBody").GetComponent<DeathRewards>();
-            deathRewards.bossPickup = new SerializablePickupIndex() { pickupName = "ItemIndex.SiphonOnLowHealth" };
-            deathRewards = LegacyResourcesAPI.Load<GameObject>("prefabs/characterbodies/ClayGrenadierBody").GetComponent<DeathRewards>();
-            deathRewards.bossPickup = new SerializablePickupIndex() { pickupName = "ItemIndex.SiphonOnLowHealth" };
-
-
-            deathRewards = LegacyResourcesAPI.Load<GameObject>("prefabs/characterbodies/ImpBody").GetComponent<DeathRewards>();
-            deathRewards.bossPickup = new SerializablePickupIndex() { pickupName = "ItemIndex.BleedOnHitAndExplode" };
-
-            deathRewards = LegacyResourcesAPI.Load<GameObject>("prefabs/characterbodies/ParentBody").GetComponent<DeathRewards>();
-            deathRewards.bossPickup = new SerializablePickupIndex() { pickupName = "ItemIndex.ParentEgg" };
-            deathRewards = Addressables.LoadAssetAsync<GameObject>(key: "RoR2/DLC2/Child/ChildBody.prefab").WaitForCompletion().GetComponent<DeathRewards>();
-            deathRewards.bossPickup = new SerializablePickupIndex() { pickupName = "ItemIndex.ParentEgg" };
-
-            deathRewards = LegacyResourcesAPI.Load<GameObject>("prefabs/characterbodies/RoboBallMiniBody").GetComponent<DeathRewards>();
-            deathRewards.bossPickup = BossItemRoboBall;
-            deathRewards = LegacyResourcesAPI.Load<GameObject>("prefabs/characterbodies/VultureBody").GetComponent<DeathRewards>();
-            deathRewards.bossPickup = BossItemRoboBall;
-
-            deathRewards = LegacyResourcesAPI.Load<GameObject>("prefabs/characterbodies/LunarExploderBody").GetComponent<DeathRewards>();
-            deathRewards.bossPickup = BossItemShinyPearl;
-            deathRewards = LegacyResourcesAPI.Load<GameObject>("prefabs/characterbodies/LunarGolemBody").GetComponent<DeathRewards>();
-            deathRewards.bossPickup = BossItemShinyPearl;
-            deathRewards = LegacyResourcesAPI.Load<GameObject>("prefabs/characterbodies/LunarWispBody").GetComponent<DeathRewards>();
-            deathRewards.bossPickup = BossItemShinyPearl;
-            deathRewards = LegacyResourcesAPI.Load<GameObject>("prefabs/characterbodies/BrotherBody").GetComponent<DeathRewards>();
-            deathRewards.bossPickup = BossItemShinyPearl;
-            deathRewards = LegacyResourcesAPI.Load<GameObject>("prefabs/characterbodies/BrotherHurtBody").GetComponent<DeathRewards>();
-            deathRewards.bossPickup = BossItemShinyPearl;
-
-
-            deathRewards = LegacyResourcesAPI.Load<GameObject>("prefabs/characterbodies/MinorConstructBody").GetComponent<DeathRewards>();
-            deathRewards.bossPickup = new SerializablePickupIndex() { pickupName = "ItemIndex.MinorConstructOnKill" };
-
-
-
-            deathRewards = LegacyResourcesAPI.Load<GameObject>("prefabs/characterbodies/VoidMegaCrabBody").GetComponent<DeathRewards>();
-            deathRewards.bossPickup = BossItemVoid;
-            deathRewards = LegacyResourcesAPI.Load<GameObject>("prefabs/characterbodies/NullifierBody").GetComponent<DeathRewards>();
-            deathRewards.bossPickup = BossItemVoid;
-            deathRewards = Addressables.LoadAssetAsync<GameObject>(key: "RoR2/DLC1/VoidJailer/VoidJailerBody.prefab").WaitForCompletion().GetComponent<DeathRewards>();
-            deathRewards.bossPickup = BossItemVoid;
-            deathRewards = Addressables.LoadAssetAsync<GameObject>(key: "RoR2/DLC1/VoidBarnacle/VoidBarnacleBody.prefab").WaitForCompletion().GetComponent<DeathRewards>();
-            deathRewards.bossPickup = BossItemVoid;
-            deathRewards = Addressables.LoadAssetAsync<GameObject>(key: "RoR2/DLC1/EliteVoid/VoidInfestorBody.prefab").WaitForCompletion().GetComponent<DeathRewards>();
-            deathRewards.bossPickup = BossItemVoid;
-            deathRewards = Addressables.LoadAssetAsync<GameObject>(key: "RoR2/DLC1/VoidRaidCrab/MiniVoidRaidCrabBodyPhase1.prefab").WaitForCompletion().AddComponent<DeathRewards>();
-            deathRewards.bossPickup = BossItemVoid;
-            deathRewards = Addressables.LoadAssetAsync<GameObject>(key: "RoR2/DLC1/VoidRaidCrab/MiniVoidRaidCrabBodyPhase2.prefab").WaitForCompletion().AddComponent<DeathRewards>();
-            deathRewards.bossPickup = BossItemVoid;
-            deathRewards = Addressables.LoadAssetAsync<GameObject>(key: "RoR2/DLC1/VoidRaidCrab/MiniVoidRaidCrabBodyPhase3.prefab").WaitForCompletion().GetComponent<DeathRewards>();
-            deathRewards.bossPickup = BossItemVoid;
-
-
-            deathRewards = Addressables.LoadAssetAsync<GameObject>(key: "RoR2/DLC2/FalseSonBoss/FalseSonBossBody.prefab").WaitForCompletion().AddComponent<DeathRewards>();
-            deathRewards.bossPickup = new SerializablePickupIndex() { pickupName = "ItemIndex.TitanGoldDuringTP" };
-            deathRewards = Addressables.LoadAssetAsync<GameObject>(key: "RoR2/DLC2/FalseSonBoss/FalseSonBossBodyLunarShard.prefab").WaitForCompletion().AddComponent<DeathRewards>();
-            deathRewards.bossPickup = new SerializablePickupIndex() { pickupName = "ItemIndex.TitanGoldDuringTP" };
-            deathRewards = Addressables.LoadAssetAsync<GameObject>(key: "RoR2/DLC2/FalseSonBoss/FalseSonBossBodyBrokenLunarShard.prefab").WaitForCompletion().GetComponent<DeathRewards>();
-            deathRewards.bossPickup = new SerializablePickupIndex() { pickupName = "ItemIndex.TitanGoldDuringTP" };
-
-
+ 
             //Nah the Charged Perforator Lemurian thing is kinda dumb even if funny
             //On.RoR2.BossGroup.OnMemberDefeatedServer += BossDropChanger;
 
+            BodyCatalog.availability.CallWhenAvailable(ChangeLate);
+            On.RoR2.BossGroup.DropRewards += BossGroup_DropRewards;
+        }
+
+        private static void BossGroup_DropRewards(On.RoR2.BossGroup.orig_DropRewards orig, BossGroup self)
+        {
+            //the old Bossrewards doesn't do a DLC check
+            //So we gotta check if theres any dlc junk, then clear the list
+            if (self.bossDrops.Count > 0)
+            {
+                if (Run.instance.IsItemExpansionLocked(self.bossDrops[0].pickupIndex.pickupDef.itemIndex))
+                {
+                    self.bossDrops.Clear();
+                }
+            }
+            orig(self);
+        }
+
+        public static void ChangeLate()
+        {
+ 
+            //1
+            RoR2Content.BodyPrefabs.GolemBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.Knurl" };
+            DLC2Content.BodyPrefabs.HalcyoniteBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.Knurl" };
+            //1
+            RoR2Content.BodyPrefabs.JellyfishBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.NovaOnLowHealth" };
+            //1
+            RoR2Content.BodyPrefabs.BeetleBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.BeetleGland" };
+            RoR2Content.BodyPrefabs.BeetleGuardBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.BeetleGland" };
+            //2
+            RoR2Content.BodyPrefabs.ClayBruiserBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.SiphonOnLowHealth" };
+            DLC1Content.BodyPrefabs.ClayGrenadierBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.SiphonOnLowHealth" };
+            //3
+            RoR2Content.BodyPrefabs.WispBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.SprintWisp" };
+            RoR2Content.BodyPrefabs.GreaterWispBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.SprintWisp" };
+            //3
+            RoR2Content.BodyPrefabs.LemurianBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.FireballsOnHit" };
+            RoR2Content.BodyPrefabs.LemurianBruiserBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.FireballsOnHit" };
+            DLC2Content.BodyPrefabs.ScorchlingBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.FireballsOnHit" };
+            //3
+            RoR2Content.BodyPrefabs.ImpBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.BleedOnHitAndExplode" };
+            //4
+            RoR2Content.BodyPrefabs.RoboBallMiniBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.RoboBallBuddy" };
+            DLC3Content.BodyPrefabs.IronHaulerBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.RoboBallBuddy" };
+            DLC3Content.BodyPrefabs.TankerBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.RoboBallBuddy" };
+            DLC3Content.BodyPrefabs.WorkerUnitBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.RoboBallBuddy" };
+            //5
+            RoR2Content.BodyPrefabs.ParentBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.ParentEgg" };
+            DLC2Content.BodyPrefabs.ChildBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.ParentEgg" };
+            //6
+            RoR2Content.BodyPrefabs.LunarExploderBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.Pearl" };
+            RoR2Content.BodyPrefabs.LunarGolemBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.Pearl" };
+            RoR2Content.BodyPrefabs.LunarWispBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.Pearl" };
+
+            #region DLC1 Items
+            //3
+            DLC1Content.BodyPrefabs.MinorConstructBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.MinorConstructOnKill" };
+
+            RoR2Content.BodyPrefabs.NullifierBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.VoidMegaCrabItem" };
+            DLC1Content.BodyPrefabs.VoidJailerBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.VoidMegaCrabItem" };
+            DLC1Content.BodyPrefabs.VoidBarnacleBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.VoidMegaCrabItem" };
+            
+            #endregion
+
+            #region DLC3 Items
+            //3
+            DLC3Content.BodyPrefabs.ExtractorUnitBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.ExtraEquipment" };
+            DLC3Content.BodyPrefabs.DefectiveUnitBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.ExtraEquipment" };
+            DLC3Content.BodyPrefabs.MinePodBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.ExtraEquipment" };
+
+            RoR2Content.BodyPrefabs.VultureBody.GetComponent<DeathRewards>().bossPickup = new SerializablePickupIndex { pickupName = "ItemIndex.ShockDamageAura" };
+
+            #endregion
         }
 
         public static void ModBossDropChanger()